+++ src/utils/getUrlParam/index.ts
@@ -1,0 +1,3 @@
+import getUrlParam from './getUrlParam';
+
+export default getUrlParam;